Add virtualenv, VCS, and other info to fish prompt
Differences from the default fish prompt include:

* shows currently active virtualenv
* only prepend (user@host) if using SSH to connect to remote system
* if PWD is a git/hg repo, show repo status in prompt

In order to minimize latency when including VCS info in the prompt,
vcprompt is used to determine whether the current directory is a
git or hg repository. (vcprompt is written in C and is extremely fast.)
If so, the appropriate VCS prompt function is called to populate that
portion of the prompt.

Why not use vcprompt's built-in info instead of using __fish_git_prompt
and hg-prompt? The latter are more configurable and flexible, and they
aren't so slow that it makes sense to rely entirely on vcprompt.