Clear line when CTRL-C is invoked in Fish shell
This was previously the default behavior in Fish, but at some point it
was changed to the current Bash-like default of appending ^C to the line
and preserving the line after canceling and moving to the next line.
This Fish configuration addition restores the original behavior of
clearing the current line when CTRL-C is invoked.